+++ canvas_modules/common-canvas/locales/command-actions/locales/zh-CN.json
