+++ canvas_modules/common-canvas/locales/common-canvas/locales/zh-CN.json
