+++ canvas_modules/common-canvas/locales/common-properties/locales/index.js
@@ -20,8 +20,8 @@ const fr = require("./fr.json");
 const itIT = require("./it.json");
 const ja = require("./ja.json");
 const ptBR = require("./pt-br.json");
-const zhTW = require("./zh-tw.json");
-const zhCN = require("./zh-cn.json");
+const zhTW = require("./zh_TW.json");
+const zhCN = require("./zh_CN.json");
 const ru = require("./ru.json");
 const eo = require("./eo.json");
 const ko = require("./ko.json");
