+++ canvas_modules/common-canvas/locales/common-properties/locales/zh-CN.json
