+++ canvas_modules/common-canvas/locales/palette/locales/zh-CN.json
