+++ canvas_modules/common-canvas/locales/toolbar/locales/zh-CN.json
